feat(#1050): ICA: context-info one-export file fragmentation
